Migrate component utils
