+++ web/html/src/components/utils/HelpIcon.tsx
@@ -1,14 +1,13 @@
-// @flow
-import * as React from 'react';
+import * as React from "react";
 
 type Props = {
-  /** Title of the icon */
-  text: ?string,
-}
+    /** Title of the icon */
+    text: string | null | undefined;
+};
 
 /** Display help icon with a title */
-const HelpIcon = ({text} : Props) : React.Node => {
+const HelpIcon = ({text}: Props) => {
     return text ? <i className="fa fa-question-circle spacewalk-help-link" title={text}></i> : null;
-}
+};
 
 export default HelpIcon;
