--- web/html/src/components/utils/StatePersistedContext.js
@@ -1,11 +1,0 @@
-/* eslint-disable */
-"use strict";
-
-import PropTypes from 'prop-types';
-import * as React from 'react';
-
-const StatePersistedContext = React.createContext({loadState: undefined, saveState: undefined});
-
-export {
-    StatePersistedContext,
-};
