+++ web/html/src/components/utils/StatePersistedContext.ts
@@ -1,0 +1,8 @@
+"use strict";
+
+import PropTypes from "prop-types";
+import * as React from "react";
+
+const StatePersistedContext = React.createContext({loadState: undefined, saveState: undefined});
+
+export {StatePersistedContext};
